# Optimize BitUtils::forEachBit