--- a/velox/common/base/CMakeLists.txt
+++ b/velox/common/base/CMakeLists.txt
@@ -27,6 +27,6 @@ add_library(velox_common_base BitUtil.cpp RawVector.cpp SimdUtil.cpp)
 target_link_libraries(velox_common_base velox_exception velox_process)
 
 add_subdirectory(test_utils)
-if(${VELOX_ENABLE_TESTING})
+if(${VELOX_BUILD_TESTING})
   add_subdirectory(benchmarks)
 endif()