Currently large SelectivityVectors, say with 1k+ rows (e.g element vector of an array vector) are slower when they have any one row not selected vs all rows selected. This is because even if one row is not selected we default to slowest way of iterating through the vector. Typically this results in the iteration being 100% slower. With this fix to forEachBit, applyToSelected() runs faster but also more consistent. This will also gracefully degrade to slowest way of iterating through vector based on number of unset bits. 


```
----------------------------------------------------------------------------
BM_isAllSelectedTrue(1000)                                 315.71us    3.17K
BM_isAllSelectedFalseLastBit(1000)                58.82%   536.78us    1.86K
BM_isAllSelectedFalseFirstBit(1000)               58.79%   537.06us    1.86K
----------------------------------------------------------------------------
============================================================================
```
Heres new :
```
----------------------------------------------------------------------------
BM_isAllSelectedTrue(1000)                                 316.38us    3.16K
BM_isAllSelectedFalseLastBit(1000)                92.80%   340.91us    2.93K
BM_isAllSelectedFalseFirstBit(1000)               89.65%   352.89us    2.83K
----------------------------------------------------------------------------
============================================================================
```

**How it works:**
* We take one word at a time
* Check if all bits are set, if true : Then run func() over set bits using for loop . (fastest)
* Else if any one bit is unset, use while loop to individually run func() over selected bits. (slowest)
* ForEachWord, takes a partial word function, and a full word function which implement the above .