@mbasmanova 2022-02-02
> Made the changes to forEachBit and I am seeing similar numbers:

@kgpai Thank you for checking. Sounds like the optimization can be re-implemented by changing the implementation of forEachBit where it would benefit more code paths and would allow to preserve somewhat short and simple implementation for applyToSelected and testSelected.