@kgpai 2022-02-02
Update : Moved changes to forEachBit.