@kgpai 2022-02-04
- Simplified forEachBit so we only use fast case when all bits are set
- Added Test case 
- Added benchmark , benchmark results are in description.